+++ _third_party/gopkg.in/fsnotify.v1/fsnotify_bsd.go
@@ -17,71 +17,56 @@ import (
 )
 
 const (
-	// Flags (from <sys/event.h>)
-	sys_NOTE_DELETE = 0x0001 /* vnode was removed */
-	sys_NOTE_WRITE  = 0x0002 /* data contents changed */
-	sys_NOTE_EXTEND = 0x0004 /* size increased */
-	sys_NOTE_ATTRIB = 0x0008 /* attributes changed */
-	sys_NOTE_LINK   = 0x0010 /* link count changed */
-	sys_NOTE_RENAME = 0x0020 /* vnode was renamed */
-	sys_NOTE_REVOKE = 0x0040 /* vnode access was revoked */
-
-	// Watch all events
-	sys_NOTE_ALLEVENTS = sys_NOTE_DELETE | sys_NOTE_WRITE | sys_NOTE_ATTRIB | sys_NOTE_RENAME
+	// Watch all events (except NOTE_EXTEND, NOTE_LINK, NOTE_REVOKE)
+	noteAllEvents = syscall.NOTE_DELETE | syscall.NOTE_WRITE | syscall.NOTE_ATTRIB | syscall.NOTE_RENAME
 
 	// Block for 100 ms on each call to kevent
 	keventWaitTime = 100e6
 )
 
-type FileEvent struct {
-	mask   uint32 // Mask of events
-	Name   string // File name (optional)
-	create bool   // set by fsnotify package if found new file
+// newEvent returns an platform-independent Event based on kqueue Fflags.
+func newEvent(name string, mask uint32, create bool) Event {
+	e := Event{Name: name}
+	if create {
+		e.Op |= Create
+	}
+	if mask&syscall.NOTE_DELETE == syscall.NOTE_DELETE {
+		e.Op |= Remove
+	}
+	if mask&syscall.NOTE_WRITE == syscall.NOTE_WRITE {
+		e.Op |= Write
+	}
+	if mask&syscall.NOTE_RENAME == syscall.NOTE_RENAME {
+		e.Op |= Rename
+	}
+	if mask&syscall.NOTE_ATTRIB == syscall.NOTE_ATTRIB {
+		e.Op |= Chmod
+	}
+	return e
 }
 
-// IsCreate reports whether the FileEvent was triggered by a creation
-func (e *FileEvent) IsCreate() bool { return e.create }
-
-// IsDelete reports whether the FileEvent was triggered by a delete
-func (e *FileEvent) IsDelete() bool { return (e.mask & sys_NOTE_DELETE) == sys_NOTE_DELETE }
-
-// IsModify reports whether the FileEvent was triggered by a file modification
-func (e *FileEvent) IsModify() bool {
-	return ((e.mask&sys_NOTE_WRITE) == sys_NOTE_WRITE || (e.mask&sys_NOTE_ATTRIB) == sys_NOTE_ATTRIB)
-}
-
-// IsRename reports whether the FileEvent was triggered by a change name
-func (e *FileEvent) IsRename() bool { return (e.mask & sys_NOTE_RENAME) == sys_NOTE_RENAME }
-
-// IsAttrib reports whether the FileEvent was triggered by a change in the file metadata.
-func (e *FileEvent) IsAttrib() bool {
-	return (e.mask & sys_NOTE_ATTRIB) == sys_NOTE_ATTRIB
-}
-
+// Watcher watches a set of files, delivering events to a channel.
 type Watcher struct {
+	Events          chan Event
+	Errors          chan error
 	mu              sync.Mutex          // Mutex for the Watcher itself.
-	kq              int                 // File descriptor (as returned by the kqueue() syscall)
-	watches         map[string]int      // Map of watched file descriptors (key: path)
+	kq              int                 // File descriptor (as returned by the kqueue() syscall).
+	watches         map[string]int      // Map of watched file descriptors (key: path).
 	wmut            sync.Mutex          // Protects access to watches.
-	fsnFlags        map[string]uint32   // Map of watched files to flags used for filter
-	fsnmut          sync.Mutex          // Protects access to fsnFlags.
-	enFlags         map[string]uint32   // Map of watched files to evfilt note flags used in kqueue
+	enFlags         map[string]uint32   // Map of watched files to evfilt note flags used in kqueue.
 	enmut           sync.Mutex          // Protects access to enFlags.
-	paths           map[int]string      // Map of watched paths (key: watch descriptor)
-	finfo           map[int]os.FileInfo // Map of file information (isDir, isReg; key: watch descriptor)
+	paths           map[int]string      // Map of watched paths (key: watch descriptor).
+	finfo           map[int]os.FileInfo // Map of file information (isDir, isReg; key: watch descriptor).
 	pmut            sync.Mutex          // Protects access to paths and finfo.
-	fileExists      map[string]bool     // Keep track of if we know this file exists (to stop duplicate create events)
+	fileExists      map[string]bool     // Keep track of if we know this file exists (to stop duplicate create events).
 	femut           sync.Mutex          // Protects access to fileExists.
 	externalWatches map[string]bool     // Map of watches added by user of the library.
 	ewmut           sync.Mutex          // Protects access to externalWatches.
-	Error           chan error          // Errors are sent on this channel
-	internalEvent   chan *FileEvent     // Events are queued on this channel
-	Event           chan *FileEvent     // Events are returned on this channel
 	done            chan bool           // Channel for sending a "quit message" to the reader goroutine
 	isClosed        bool                // Set to true when Close() is first called
 }
 
-// NewWatcher creates and returns a new kevent instance using kqueue(2)
+// NewWatcher establishes a new watcher with the underlying OS and begins waiting for events.
 func NewWatcher() (*Watcher, error) {
 	fd, errno := syscall.Kqueue()
 	if fd == -1 {
@@ -90,26 +75,21 @@ func NewWatcher() (*Watcher, error) {
 	w := &Watcher{
 		kq:              fd,
 		watches:         make(map[string]int),
-		fsnFlags:        make(map[string]uint32),
 		enFlags:         make(map[string]uint32),
 		paths:           make(map[int]string),
 		finfo:           make(map[int]os.FileInfo),
 		fileExists:      make(map[string]bool),
 		externalWatches: make(map[string]bool),
-		internalEvent:   make(chan *FileEvent),
-		Event:           make(chan *FileEvent),
-		Error:           make(chan error),
+		Events:          make(chan Event),
+		Errors:          make(chan error),
 		done:            make(chan bool, 1),
 	}
 
 	go w.readEvents()
-	go w.purgeEvents()
 	return w, nil
 }
 
-// Close closes a kevent watcher instance
-// It sends a message to the reader goroutine to quit and removes all watches
-// associated with the kevent instance
+// Close removes all watches and closes the events channel.
 func (w *Watcher) Close() error {
 	w.mu.Lock()
 	if w.isClosed {
@@ -119,21 +99,22 @@ func (w *Watcher) Close() error {
 	w.isClosed = true
 	w.mu.Unlock()
 
-	// Send "quit" message to the reader goroutine
+	// Send "quit" message to the reader goroutine:
 	w.done <- true
 	w.wmut.Lock()
 	ws := w.watches
 	w.wmut.Unlock()
-	for path := range ws {
-		w.removeWatch(path)
+	for name := range ws {
+		w.Remove(name)
 	}
 
 	return nil
 }
 
-// AddWatch adds path to the watched file set.
+// addWatch adds path to the watched file set.
 // The flags are interpreted as described in kevent(2).
 func (w *Watcher) addWatch(path string, flags uint32) error {
+	path = filepath.Clean(path)
 	w.mu.Lock()
 	if w.isClosed {
 		w.mu.Unlock()
@@ -175,9 +156,9 @@ func (w *Watcher) addWatch(path string, flags uint32) error {
 			}
 		}
 
-		fd, errno := syscall.Open(path, open_FLAGS, 0700)
+		fd, errno := syscall.Open(path, openMode, 0700)
 		if fd == -1 {
-			return errno
+			return os.NewSyscallError("Open", errno)
 		}
 		watchfd = fd
 
@@ -194,8 +175,8 @@ func (w *Watcher) addWatch(path string, flags uint32) error {
 	w.pmut.Lock()
 	w.enmut.Lock()
 	if w.finfo[watchfd].IsDir() &&
-		(flags&sys_NOTE_WRITE) == sys_NOTE_WRITE &&
-		(!found || (w.enFlags[path]&sys_NOTE_WRITE) != sys_NOTE_WRITE) {
+		(flags&syscall.NOTE_WRITE) == syscall.NOTE_WRITE &&
+		(!found || (w.enFlags[path]&syscall.NOTE_WRITE) != syscall.NOTE_WRITE) {
 		watchDir = true
 	}
 	w.enmut.Unlock()
@@ -226,21 +207,22 @@ func (w *Watcher) addWatch(path string, flags uint32) error {
 	return nil
 }
 
-// Watch adds path to the watched file set, watching all events.
-func (w *Watcher) watch(path string) error {
+// Add starts watching the named file or directory (non-recursively).
+func (w *Watcher) Add(name string) error {
 	w.ewmut.Lock()
-	w.externalWatches[path] = true
+	w.externalWatches[name] = true
 	w.ewmut.Unlock()
-	return w.addWatch(path, sys_NOTE_ALLEVENTS)
+	return w.addWatch(name, noteAllEvents)
 }
 
-// RemoveWatch removes path from the watched file set.
-func (w *Watcher) removeWatch(path string) error {
+// Remove stops watching the the named file or directory (non-recursively).
+func (w *Watcher) Remove(name string) error {
+	name = filepath.Clean(name)
 	w.wmut.Lock()
-	watchfd, ok := w.watches[path]
+	watchfd, ok := w.watches[name]
 	w.wmut.Unlock()
 	if !ok {
-		return errors.New(fmt.Sprintf("can't remove non-existent kevent watch for: %s", path))
+		return fmt.Errorf("can't remove non-existent kevent watch for: %s", name)
 	}
 	var kbuf [1]syscall.Kevent_t
 	watchEntry := &kbuf[0]
@@ -254,10 +236,10 @@ func (w *Watcher) removeWatch(path string) error {
 	}
 	syscall.Close(watchfd)
 	w.wmut.Lock()
-	delete(w.watches, path)
+	delete(w.watches, name)
 	w.wmut.Unlock()
 	w.enmut.Lock()
-	delete(w.enFlags, path)
+	delete(w.enFlags, name)
 	w.enmut.Unlock()
 	w.pmut.Lock()
 	delete(w.paths, watchfd)
@@ -271,7 +253,7 @@ func (w *Watcher) removeWatch(path string) error {
 		w.pmut.Lock()
 		for _, wpath := range w.paths {
 			wdir, _ := filepath.Split(wpath)
-			if filepath.Clean(wdir) == filepath.Clean(path) {
+			if filepath.Clean(wdir) == filepath.Clean(name) {
 				w.ewmut.Lock()
 				if !w.externalWatches[wpath] {
 					pathsToRemove = append(pathsToRemove, wpath)
@@ -280,11 +262,11 @@ func (w *Watcher) removeWatch(path string) error {
 			}
 		}
 		w.pmut.Unlock()
-		for _, p := range pathsToRemove {
+		for _, name := range pathsToRemove {
 			// Since these are internal, not much sense in propagating error
 			// to the user, as that will just confuse them with an error about
 			// a path they did not explicitly watch themselves.
-			w.removeWatch(p)
+			w.Remove(name)
 		}
 	}
 
@@ -292,16 +274,16 @@ func (w *Watcher) removeWatch(path string) error {
 }
 
 // readEvents reads from the kqueue file descriptor, converts the
-// received events into Event objects and sends them via the Event channel
+// received events into Event objects and sends them via the Events channel
 func (w *Watcher) readEvents() {
 	var (
-		eventbuf [10]syscall.Kevent_t // Event buffer
-		events   []syscall.Kevent_t   // Received events
-		twait    *syscall.Timespec    // Time to block waiting for events
-		n        int                  // Number of events returned from kevent
-		errno    error                // Syscall errno
+		keventbuf [10]syscall.Kevent_t // Event buffer
+		kevents   []syscall.Kevent_t   // Received events
+		twait     *syscall.Timespec    // Time to block waiting for events
+		n         int                  // Number of events returned from kevent
+		errno     error                // Syscall errno
 	)
-	events = eventbuf[0:0]
+	kevents = keventbuf[0:0]
 	twait = new(syscall.Timespec)
 	*twait = syscall.NsecToTimespec(keventWaitTime)
 
@@ -317,75 +299,77 @@ func (w *Watcher) readEvents() {
 		if done {
 			errno := syscall.Close(w.kq)
 			if errno != nil {
-				w.Error <- os.NewSyscallError("close", errno)
+				w.Errors <- os.NewSyscallError("close", errno)
 			}
-			close(w.internalEvent)
-			close(w.Error)
+			close(w.Events)
+			close(w.Errors)
 			return
 		}
 
 		// Get new events
-		if len(events) == 0 {
-			n, errno = syscall.Kevent(w.kq, nil, eventbuf[:], twait)
+		if len(kevents) == 0 {
+			n, errno = syscall.Kevent(w.kq, nil, keventbuf[:], twait)
 
 			// EINTR is okay, basically the syscall was interrupted before
 			// timeout expired.
 			if errno != nil && errno != syscall.EINTR {
-				w.Error <- os.NewSyscallError("kevent", errno)
+				w.Errors <- os.NewSyscallError("kevent", errno)
 				continue
 			}
 
 			// Received some events
 			if n > 0 {
-				events = eventbuf[0:n]
+				kevents = keventbuf[0:n]
 			}
 		}
 
-		// Flush the events we received to the events channel
-		for len(events) > 0 {
-			fileEvent := new(FileEvent)
-			watchEvent := &events[0]
-			fileEvent.mask = uint32(watchEvent.Fflags)
+		// Flush the events we received to the Events channel
+		for len(kevents) > 0 {
+			watchEvent := &kevents[0]
+			mask := uint32(watchEvent.Fflags)
 			w.pmut.Lock()
-			fileEvent.Name = w.paths[int(watchEvent.Ident)]
+			name := w.paths[int(watchEvent.Ident)]
 			fileInfo := w.finfo[int(watchEvent.Ident)]
 			w.pmut.Unlock()
-			if fileInfo != nil && fileInfo.IsDir() && !fileEvent.IsDelete() {
+
+			event := newEvent(name, mask, false)
+
+			if fileInfo != nil && fileInfo.IsDir() && !(event.Op&Remove == Remove) {
 				// Double check to make sure the directory exist. This can happen when
 				// we do a rm -fr on a recursively watched folders and we receive a
 				// modification event first but the folder has been deleted and later
 				// receive the delete event
-				if _, err := os.Lstat(fileEvent.Name); os.IsNotExist(err) {
+				if _, err := os.Lstat(event.Name); os.IsNotExist(err) {
 					// mark is as delete event
-					fileEvent.mask |= sys_NOTE_DELETE
+					event.Op |= Remove
 				}
 			}
 
-			if fileInfo != nil && fileInfo.IsDir() && fileEvent.IsModify() && !fileEvent.IsDelete() {
-				w.sendDirectoryChangeEvents(fileEvent.Name)
+			if fileInfo != nil && fileInfo.IsDir() && event.Op&Write == Write && !(event.Op&Remove == Remove) {
+				w.sendDirectoryChangeEvents(event.Name)
 			} else {
-				// Send the event on the events channel
-				w.internalEvent <- fileEvent
+				// Send the event on the Events channel
+				w.Events <- event
 			}
 
 			// Move to next event
-			events = events[1:]
+			kevents = kevents[1:]
 
-			if fileEvent.IsRename() {
-				w.removeWatch(fileEvent.Name)
+			if event.Op&Rename == Rename {
+				w.Remove(event.Name)
 				w.femut.Lock()
-				delete(w.fileExists, fileEvent.Name)
+				delete(w.fileExists, event.Name)
 				w.femut.Unlock()
 			}
-			if fileEvent.IsDelete() {
-				w.removeWatch(fileEvent.Name)
+			if event.Op&Remove == Remove {
+				w.Remove(event.Name)
 				w.femut.Lock()
-				delete(w.fileExists, fileEvent.Name)
+				delete(w.fileExists, event.Name)
 				w.femut.Unlock()
 
 				// Look for a file that may have overwritten this
 				// (ie mv f1 f2 will delete f2 then create f2)
-				fileDir, _ := filepath.Split(fileEvent.Name)
+				fileDir, _ := filepath.Split(event.Name)
 				fileDir = filepath.Clean(fileDir)
 				w.wmut.Lock()
 				_, found := w.watches[fileDir]
@@ -415,18 +399,9 @@ func (w *Watcher) watchDirectoryFiles(dirPath string) error {
 	for _, fileInfo := range files {
 		filePath := filepath.Join(dirPath, fileInfo.Name())
 
-		// Inherit fsnFlags from parent directory
-		w.fsnmut.Lock()
-		if flags, found := w.fsnFlags[dirPath]; found {
-			w.fsnFlags[filePath] = flags
-		} else {
-			w.fsnFlags[filePath] = FSN_ALL
-		}
-		w.fsnmut.Unlock()
-
 		if fileInfo.IsDir() == false {
 			// Watch file to mimic linux fsnotify
-			e := w.addWatch(filePath, sys_NOTE_ALLEVENTS)
+			e := w.addWatch(filePath, noteAllEvents)
 			if e != nil {
 				return e
 			}
@@ -436,7 +411,7 @@ func (w *Watcher) watchDirectoryFiles(dirPath string) error {
 			w.enmut.Lock()
 			currFlags, found := w.enFlags[filePath]
 			w.enmut.Unlock()
-			var newFlags uint32 = sys_NOTE_DELETE
+			var newFlags uint32 = syscall.NOTE_DELETE
 			if found {
 				newFlags |= currFlags
 			}
@@ -463,7 +438,7 @@ func (w *Watcher) sendDirectoryChangeEvents(dirPath string) {
 	// Get all files
 	files, err := ioutil.ReadDir(dirPath)
 	if err != nil {
-		w.Error <- err
+		w.Errors <- err
 	}
 
 	// Search for new files
@@ -473,24 +448,32 @@ func (w *Watcher) sendDirectoryChangeEvents(dirPath string) {
 		_, doesExist := w.fileExists[filePath]
 		w.femut.Unlock()
 		if !doesExist {
-			// Inherit fsnFlags from parent directory
-			w.fsnmut.Lock()
-			if flags, found := w.fsnFlags[dirPath]; found {
-				w.fsnFlags[filePath] = flags
-			} else {
-				w.fsnFlags[filePath] = FSN_ALL
-			}
-			w.fsnmut.Unlock()
-
-			// Send create event
-			fileEvent := new(FileEvent)
-			fileEvent.Name = filePath
-			fileEvent.create = true
-			w.internalEvent <- fileEvent
+			// Send create event (mask=0)
+			event := newEvent(filePath, 0, true)
+			w.Events <- event
 		}
+
+		// watchDirectoryFiles (but without doing another ReadDir)
+		if fileInfo.IsDir() == false {
+			// Watch file to mimic linux fsnotify
+			w.addWatch(filePath, noteAllEvents)
+		} else {
+			// If the user is currently watching directory
+			// we want to preserve the flags used
+			w.enmut.Lock()
+			currFlags, found := w.enFlags[filePath]
+			w.enmut.Unlock()
+			var newFlags uint32 = syscall.NOTE_DELETE
+			if found {
+				newFlags |= currFlags
+			}
+
+			// Linux gives deletes if not explicitly watching
+			w.addWatch(filePath, newFlags)
+		}
+
 		w.femut.Lock()
 		w.fileExists[filePath] = true
 		w.femut.Unlock()
 	}
-	w.watchDirectoryFiles(dirPath)
 }